+++ magda-search-api/src/main/scala/au/csiro/data61/magda/search/elasticsearch/Queries.scala
@@ -19,16 +19,25 @@ import au.csiro.data61.magda.model.misc.Region
 import au.csiro.data61.magda.search.SearchStrategy
 
 object Queries {
-  def publisherQuery(strategy: SearchStrategy)(publisher: FilterValue[String]) = {
+  def publisherQuery(strategy: SearchStrategy, searchForAcronym: Boolean = false)(publisher: FilterValue[String]) = {
     handleFilterValue(publisher, (publisherString: String) =>
       strategy match {
-        case SearchStrategy.MatchAll => multiMatchQuery(publisherString)
-          .fields("publisher.acronym", "publisher.name.keyword_lowercase")
-          .minimumShouldMatch("-50%")
-        case SearchStrategy.MatchPart =>
-          multiMatchQuery(publisherString)
-            .fields("publisher.acronym", "publisher.name", "publisher.name.english")
-            .minimumShouldMatch("-50%")
+        case SearchStrategy.MatchAll => if(!searchForAcronym) {
+            matchQuery("publisher.name.keyword_lowercase", publisherString)
+        } else { 
+            multiMatchQuery(publisherString)
+                .fields("publisher.acronym", "publisher.name.keyword_lowercase")
+                .minimumShouldMatch("-50%")
+        }
+        case SearchStrategy.MatchPart => if(!searchForAcronym) {
+            multiMatchQuery(publisherString)
+                .fields("publisher.name", "publisher.name.english")
+                .minimumShouldMatch("-50%")
+        } else {
+            multiMatchQuery(publisherString)
+                .fields("publisher.acronym", "publisher.name", "publisher.name.english")
+                .minimumShouldMatch("-50%")
+        }
       }, "publisher.name"
     )
   }